Node::getParent(): throw exception to convey NULL semantics.
C++ is a statically typed language, but Python is not. It is therefore
tricky to convey the same semantics - in which a NULL would mean a node
with no parent - to Python code without getting into optionals (C++17)
and changing the binding API.

This solution compromises by throwing an exception instead, which gets
converted to a Python exception automatically by pybind11. Given this,
the calling Python code can catch the exception and break any loops if
applicable and the C++ code will not create objects with no underlying
RAMSES pointers, which are useless.

Signed-off-by: Daniel W. S. Almeida <dwlsalmeida@gmail.com>